# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/learning/courses/{slug}": {
      "get": {
        "operationId": "LearningController_getCourse",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/lessons/{slug}": {
      "get": {
        "operationId": "LearningController_getLesson",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/lessons/{id}/progress": {
      "patch": {
        "operationId": "LearningController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/courses/{id}/progress": {
      "get": {
        "operationId": "LearningController_myProgress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/recommendations/report/{reportId}": {
      "get": {
        "operationId": "LearningController_recommendations",
        "parameters": [
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/admin/learning/courses": {
      "get": {
        "operationId": "LearningAdminController_listCourses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      },
      "post": {
        "operationId": "LearningAdminController_createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      }
    },
    "/api/v1/admin/learning/courses/{id}": {
      "get": {
        "operationId": "LearningAdminController_getCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      },
      "patch": {
        "operationId": "LearningAdminController_updateCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      },
      "delete": {
        "operationId": "LearningAdminController_archiveCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      }
    },
...
    "/api/v1/admin/learning/lessons/{id}": {
      "get": {
        "operationId": "LearningAdminController_getLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      },
      "patch": {
        "operationId": "LearningAdminController_updateLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      },
      "delete": {
        "operationId": "LearningAdminController_archiveLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning Admin"
        ]
      }
    },
...
    "/api/v1/admin/gamification/badges": {
      "get": {
        "operationId": "GamificationAdminController_listBadges",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gamification Admin"
        ]
      },
      "post": {
        "operationId": "GamificationAdminController_createBadge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBadgeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gamification Admin"
        ]
      }
    },
    "/api/v1/admin/gamification/badges/{id}": {
      "get": {
        "operationId": "GamificationAdminController_getBadge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gamification Admin"
        ]
      },
      "patch": {
        "operationId": "GamificationAdminController_updateBadge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBadgeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gamification Admin"
        ]
      },
      "delete": {
        "operationId": "GamificationAdminController_archiveBadge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Gamification Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Verith API",
    "description": "Explainable misinformation verification and media literacy API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "verith_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "verith_refresh"
      }
    },
    "schemas": {
...
      "AdminReasonDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateLessonProgressDto": {
        "type": "object",
        "properties": {
          "progress": {
            "type": "number"
          },
          "lastPosition": {
            "type": "number"
          }
        },
        "required": [
          "progress"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateCourseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLessonDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBadgeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBadgeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}